backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database (SQLite for MVP)
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI/ML
google-generativeai==0.8.3
pillow==10.2.0
rembg==2.0.50

# HTTP client
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
